add ability to resend request if it isn't fit in reply buffer. init llog too early.
Branch b1_8
b=19526
i=nathan
i=e.mei
i=tappro

+++ lustre/mds/mds_reint.c
@@ -1834,7 +1834,7 @@ void mds_shrink_body_reply(struct ptlrpc_request *req,
                            int reply_mdoff)
 {
         struct mds_body *rq_body;
-        const long have_acl = OBD_MD_FLCOOKIE | OBD_MD_FLACL;
+        const long long have_acl = OBD_MD_FLCOOKIE | OBD_MD_FLACL;
         const long have_md = OBD_MD_FLEASIZE | OBD_MD_FLDIREA;
         ENTRY;
